Increase UV_THREADPOOL_SIZE to 64
This threadpool is used by Node to allow synchronous tasks to be
performed asynchronously and not block the main node thread.

Tasks performed by this threadpool include, but are not limited to,
 DNS resolution -
see https://nodejs.org/api/cli.html#cli_uv_threadpool_size_size

The threadpool can become a bottleneck if eg. DNS requests are slow or
timeout, leading to requests queueing as they wait for an available
worker thread.

Increasing the size of this threadpool is desirable for the traffic
profile of api gateway, which typically handles a large number of
concurrent requests with requests typically I/O rather than CPU
bound.