CCD-2460: Remove redundant integration test and associated files
- Removed root level integration_test directory and contents (pom.xml, log4j.properties and ApiGatewayIT.java).
- Removed test:acceptance line from scripts section of package.json which only output the message 'TODO - CALL JAVA INTEGRATION TEST'

The directory and contents were part of original commit to GitHub in March 2018 and hadn't been changed since.  Test was not being invoked by any pipeline and only reference to it in the code was a the TODO line in package.json.  Evidence suggests that the test has not been run since 2018 and is therefore redundant.

Removing the pom.xml file also addresses a number of CVEs that were detected in it by Dependabot scan.